remove comments

+++ ft_memset.c
@@ -4,7 +4,7 @@
 void *ft_memset(void *dest, int c, size_t count)
 {
     unsigned char *d = (unsigned char*)dest;
-    int i;
+    size_t i;
     i = 0;
     while (i < count)
         d[i++] = (unsigned char)c;
@@ -15,7 +15,7 @@ int main(void)
 {
    char buffer[21];
    void *string;
- 
+
    ft_memset(buffer, 0, sizeof(buffer));
    string = (char *) ft_memset(buffer,'P', 10);
    printf("\nBuffer contents: %s\n", (char*)string);
